my-prism-dapp/requirements.txt
requests==2.31.0
Pillow==9.3.0
python-dotenv==1.0.0
